# contracts/openapi/platform-draft.yaml
openapi: 3.1.0
info:
  title: OpsWeave platform API
  version: 0.1.0-draft
  description: >
    Identity-scoped business endpoints. Tenant and permissions are taken from the
    trusted principal, never from query or body. Explicit OIDC mode uses the authorization-code BFF;
    dev mode accepts only its configured loopback credential. Cookie requests require exact Origin
    and X-CSRF-TOKEN on unsafe methods; GET session metadata supplies the masked token.
    Zabbix Host ingest is available only when an explicit source mode is configured.
    All API responses use Cache-Control no-store and an X-OpsWeave-Request-Id UUID.
    The optional client request ID is correlation metadata only, not identity or idempotency.
paths:
...
  /api/v1/integrations/workflows/runtime:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
    get:
      operationId: getWorkflowRuntime
      description: Read owner-scoped runtime tasks and executions. Runtime mode and background availability are server-derived and never grant authority.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Runtime status
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-runtime.schema.json'}
        default: {description: 401 unauthenticated; 403 denied; 503 runtime unavailable}
  /api/v1/integrations/workflows/runtime/host-scans/{id}:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, pattern: '^[a-z][a-z0-9_-]{0,47}$'}}
    get:
      operationId: getWorkflowRuntimeHostScan
      description: Read the bounded checkpoint and confirmed batches for one owner-scoped runtime host scan.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Runtime host scan state
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-host-scan.schema.json'}
        default: {description: 401 unauthenticated; 403 denied; 404 hidden scan; 503 runtime unavailable}
  /api/v1/integrations/workflows/runtime/commands/{requestId}:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: requestId, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: getWorkflowRuntimeCommandReceipt
      description: Read the original owner-bound runtime control receipt after an uncertain response. It never repeats the command.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Runtime control receipt
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-runtime-control-receipt.schema.json'}
        default: {description: 401 unauthenticated; 403 denied; 404 hidden receipt; 503 runtime unavailable}
  /api/v1/integrations/workflows/runtime/execute:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
    post:
      operationId: executeWorkflowRuntime
      description: Execute a pinned workflow against a bounded explicit sample or completed source run under server-side identity, scope, and budget checks.
      security: [{bearerAuth: []}, {browserSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '../schemas/v2/workflow-runtime-execute-request.schema.json'}
      responses:
        '200':
          description: Runtime execution
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-runtime-execution.schema.json'}
        default: {description: 400 invalid closed input; 401 unauthenticated; 403 denied; 404 hidden workflow; 409 stale pin; 503 runtime unavailable}
...
components:
  parameters:
    clientRequestId:
      in: header
      name: X-OpsWeave-Request-Id
      required: false
      description: Optional single UUID; malformed or duplicate values return 400. Server generates one if absent and echoes a canonical UUID on every API response. Never grants permissions or suppresses duplicate writes.
      schema: {type: string, format: uuid}
      example: 6c8d4385-c206-4b59-8c2e-f9d34f8b3500
...
  securitySchemes:
    historyService:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Dedicated OAuth client_credentials access token profile described in ../history-service-identity.md. Never accepted as a browser or Runtime credential.
    runtimeKey:
      type: apiKey
      in: header
      name: X-OpsWeave-Runtime-Key
      description: Explicit local process attestation only; does not grant user permissions and is never exposed to Web/model.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Explicit loopback dev credential, or short-lived server-only Runtime delegation on allowlisted endpoints. Raw provider bearer JWTs are not accepted.
    browserSession:
      type: apiKey
      in: cookie
      name: __Host-opsweave
      description: OIDC BFF HttpOnly Secure SameSite=Lax host cookie. Explicit loopback protocol tests use opsweave-test-session. Unsafe methods also require X-CSRF-TOKEN and the configured Origin.